% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption.check.R
\name{assumption.check}
\alias{assumption.check}
\title{Check assumption for parametrics tests}
\usage{
assumption.check(x, g1, g2 = NULL, ...)
}
\arguments{
\item{x}{Response variable required to be analyzed}

\item{g1}{Required explanatory variable (first predictor)}

\item{g2}{Optional explanatory variable (second predictor)}
}
\value{
Shapiro–Wilk test result and Bartlett's Test result with graphic checks
}
\description{
This function will return the p values of  Shapiro–Wilk test
result and Bartlett's Test result. At the same time, this fucntion will output
residual plot with QQ plot by package ggfortify. It will also write suggestion
about normality and equality of variance of the dataset. User can maximally
imput two factors.
}
\examples{
assumption.check(poisons$time, poisons$treat)
assumption.check(poisons$time, poisons$treat, poisons$poison)
}
\keyword{groups}
