% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Bar chart visualiztion of the data summary}
\usage{
barchart(y, x1, x2 = NULL, normal = T, edgecol = 1,
               fillcol1 = rep(5:(length(unique(x1)) + 4)), 
               fillcol2 = rep(5:(length(unique(x2)) + 4)), 
               main = "BAR CHART", titlesize = 13, titleface = "bold", 
               titlefont = 'Times New Roman', titlecol = 1, 
               titlehjust = 0.5, subtitle = NULL, subsize = NULL,
               subface = NULL, subfont = NULL, subcol = 1, 
               subhjust = 0.5, caption = NULL, capsize = NULL, 
               capfont = NULL, capcol = "gray48", caphjust = 0, 
               legname = NULL, leglab = unique(x2), legsize = 11,
               legcol = 1, legface = NULL, legfont = NULL, legtxsize = 10,
               legtxcol = 1, legtxface = NULL, legtxfont = NULL, 
               legpos = "right", psize = 4, pvalpos = min(y, na.rm = T), 
               pfont = 'Times New Roman',globalp_pos = max(y, na.rm = T), 
               na.rm = T, xname = NULL, xnamesize = 12, xnamecol = 1, 
               xnameface = 'bold', xnamefont = 'Times New Roman', 
               xlabels = unique(x1), xlabelsize = 10, xlabelcol = "grey48", 
               xlabelface = NULL, xlabelfont = 'Arial', yname = NULL, 
               ynamesize = 12, ynamecol = 1, ynameface = 'bold', 
               ynamefont = 'Times New Roman', ylabelsize = NULL, 
               ylabelcol = "grey48", ylabelface = NULL, ylabelfont = 'Arial', .
               na.rm = T, ...)
}
\arguments{
\item{y}{Response variable required to be analyzed}

\item{x1}{Required explanatory variable (first predictor)}

\item{x2}{Optional explanatory variable (second predictor)}

\item{normal}{ANOVA test will be procesed if TRUE or Kruskal-Waills test will be processed if FALSE}

\item{edgecol}{The egde color of the boxplots (default is black)}

\item{fillcol1}{The fill-in color of the boxplots when SINGLE predictor (default is white)}

\item{fillcol2}{The fill-in color of the boxplots when TWO predictors (default depends on the total number of levels)}

\item{main}{Title of the boxplot}

\item{titlesize}{Font size of title}

\item{titleface}{Font style of title (e.g. 'bold')}

\item{titlefont}{Font of title (e.g. 'Times New Roman')}

\item{titlecol}{Color of title (default is black)}

\item{titlehjust}{Position of title (default is middle)}

\item{subtitle}{Subtitle of the boxplot}

\item{subsize}{Font size of subtitle}

\item{subface}{Font style of the subtitle}

\item{subfont}{Font of subtitle}

\item{subcol}{Color of subtitle (default is black)}

\item{subhjust}{Position of subtitle (default is middle)}

\item{caption}{Caption of the boxplot}

\item{capsize}{Font size of caption}

\item{capfont}{Font of caption}

\item{capcol}{Color of caption (default is 'gray48')}

\item{caphjust}{Position of caption (default is left)}

\item{legname}{Name of legend when TWO predictors}

\item{legsize}{Font size of legend title}

\item{legcol}{Color of legend title (default is black)}

\item{legface}{Font style of legend title}

\item{legfont}{Font of legend title}

\item{leglab}{Labels of each level in legend}

\item{legtxsize}{Font size of legend levels}

\item{legtxcol}{Font color of legend levels}

\item{legtxface}{Font style of legend levels}

\item{legtxfont}{Font of legend levels}

\item{legpos}{Position of legend (default is right)}

\item{xname}{Name of the variable on the x-asis}

\item{xnamesize}{Font size of the variable on the x-asis}

\item{xnamecol}{Font color of the variable on the x-asis}

\item{xnameface}{Font style of the variable on the x-asis}

\item{xnamefont}{Font of the variable on the x-asis}

\item{xlabels}{Labels of the levels on the x-asis}

\item{xlabelsize}{Font size of the levels on the x-asis}

\item{xlabelcol}{Font color of the levels on the x-asis}

\item{xlabelface}{Font style of the levels on the x-asis}

\item{xlabelfont}{Font of the levels on the x-asis}

\item{yname}{Name of the variable on the y-asis}

\item{ynamesize}{Font size of the variable on the y-asis}

\item{ynamecol}{Font color of the variable on the y-asis}

\item{ynameface}{Font style of the variable on the y-asis}

\item{ynamefont}{Font of the variable on the y-asis}

\item{ylabelsize}{Labels of the levels on the y-asis}

\item{ylabelcol}{Font color of the levels on the y-asis}

\item{ylabelface}{Font style of the levels on the y-asis}

\item{ylabelfont}{Font of the levels on the y-asis}

\item{psize}{Text size of p-value(s)}

\item{pvalpos}{Position of p-values for TWO predictors (default is at the bottom)}

\item{pfont}{Font of p-values}

\item{globalp_pos}{Position of global p-value}

\item{na.rm}{Whether to include missing values (default is TRUE)}

\item{...}{Further arguments to be passed to or from methods.}
}
\description{
Summarize continuous response variable by categorical explanatory variable(s) with bar chart(s)
}
\details{
A Bar Chart Function

The barchart function return the barchart from ggplot2. Response variable should be continuous. 
         Predictor(s) can be ONE or TWO, but all required to be categorical variable(s). Levels more than 
         10 are not recommended for this function in order to have good visualization.
}
\keyword{barchart}
\keyword{comparison}
\keyword{summary}
