% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxTrans.R
\name{boxcoxTrans}
\alias{boxcoxTrans}
\title{Data transformation}
\usage{
boxcoxTrans(x, g1, g2=NULL, ...)
}
\arguments{
\item{x}{Response variable required to be analyzed}

\item{g1}{Required explanatory variable (first predictor)}

\item{g2}{Optional explanatory variable (second predictor)}

\item{lb}{plot start at this x value}

\item{ub}{plot end at this x value}

\item{bk}{the interval of x values in the plot}

\item{plotit}{logical which controls whether the result should be plotted.}
}
\value{
table (data.frame) with adjusted resonse variable
}
\description{
Transform continous data by Box-Cox Transformation
}
\details{
This function return a data frame with one more colume of the transformed reponse variable.
This function automatically take the imputs varible names. For example, if the impute is x=data$nameA,
the this function will know the name of varible x is nameA. The transformed data name will be
nameA.trans in the output data frame.
}
\examples{
boxcoxTrans(poisons$time, poisons$treat)
boxcoxTrans(poisons$time, poisons$treat, poisons$poison)
}
\keyword{transformation}
