% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.2.factors}
\alias{compare.2.factors}
\title{Compare four groups of continous varibles by two factors}
\usage{
compare.2.factors(x, g, holm = TRUE, bonferroni = FALSE, ...)
}
\arguments{
\item{x}{Response variable required to be analyzed}

\item{g}{Required explanatory variable}

\item{holm}{a logical indicating whether you want show Holm-Bonferroni adjusted p.values for t.test.
Defaults to FALSE.}

\item{bonferroni}{a logical indicating whether you want show Bonferroni adjusted p.values for t.test.
Defaults to FALSE.}
}
\value{
table of p values
}
\description{
This function will write a summary tables of p values and adjusted
 p values to show the main effect of the two factors and the interaction between the two factors.
}
\examples{
compare.2.factors(poisons$time, poisons$treat)
}
\keyword{groups}
