% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupsum.R
\name{groupsum}
\alias{groupsum}
\title{Compare continuous data based on groups}
\usage{
groups.summary(x, g1, g2 = NULL, ...)
}
\arguments{
\item{x}{Response variable required to be analyzed}

\item{g1}{Required explanatory variable (first predictor)}

\item{g2}{Optional explanatory variable (second predictor)}
}
\value{
table summary for groups of continous variables
}
\description{
This function will write a summary tables for continous variables by groups.
This function can maximum take two factors to write out a summary table. The summary table
has the number of varibales, the missing values, mean, standard deviation, median and
quantile for all groups. If there's one factor, table will show the p-values of ANOVA
and Kruskal-Wallis test at the last row. If there are two factors, the table will only show
the global p-value from the ANOVA test.
}
\examples{
groupsum(poisons$time,poisons$treat)
groupsum(poisons$time,poisons$treat, poisons$poison)
}
\keyword{groups}
