% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{p.adj.compare}
\alias{p.adj.compare}
\title{Adjusted p values for pairwise comparisons}
\usage{
p.adj.compare(x, g, control = levels(g)[1], level.names = levels(g),
                     n = match(control,level.names),p.digit = 3, paired = FALSE,
                     compareall = FALSE, t.test = TRUE, wilcoxon = TRUE,
                      wilcoxon.holm = TRUE, holm = TRUE, bonferroni = FALSE,
                      Dunnett = TRUE, hochberg = TRUE, hommel=TRUE,
                      BH = TRUE, BY = TRUE, Tukey = TRUE, ...)
}
\arguments{
\item{x}{Response variable required to be analyzed.}

\item{g}{Required explanatory variable (grouping vector or factor).}

\item{control}{indicate one group to compare with all other groups. Defaults to first group.}

\item{p.digit}{the decimal places for p.values. Defaults to 3.}

\item{paired}{a logical indicating whether you want a paired test. Defaults to FALSE.}

\item{compareall}{if TRUE do pairwise comparsion for all groups and return p values. Defaults to FALSE.}

\item{t.test}{a logical indicating whether you want to do two sample t.test result p.values without
adjustment. Defaults to TRUE. If FALSE, then this function can only show Wilcoxon tests result p.values
and Holm-Bonferroni adjusted p.values for Wilcoxon tests.}

\item{wilcoxon}{a logical indicating whether you want show two sample Wilcoxon tests result p.values
without adjustment. Defaults to TRUE.}

\item{wilcoxon.holm}{a logical indicating whether you want show Holm-Bonferroni adjusted p.values for
Wilcoxon tests.}

\item{holm}{a logical indicating whether you want show Holm-Bonferroni adjusted p.values for t.test.
Defaults to FALSE.}

\item{hochberg}{a logical indicating whether you want show Hochberg's adjusted p.values for t.test.
Defaults to FALSE.}

\item{hommel}{a logical indicating whether you want show Hommel's adjusted p.values for t.test.
Defaults to FALSE.}

\item{BH}{a logical indicating whether you want show Benjamini adjusted p.values for t.test.
Defaults to FALSE.}

\item{BY}{a logical indicating whether you want show Yekutieli adjusted p.values for t.test.
Defaults to FALSE.}

\item{bonferroni}{a logical indicating whether you want show Bonferroni adjusted p.values for t.test.
Defaults to FALSE.}

\item{dunnett}{a logical indicating whether you want show Dunnett's Test adjusted p.values. Defaults
to TRUE.}

\item{tukey}{a logical indicating whether you want show Tukey's test adjusted p.values. Defaults to TRUE.}
}
\value{
table (data.frame) of p values
}
\description{
Multiple types of adjusted p values of pairwise comparisons between group levels.
}
\details{
Only if t.test is True, this program can show adjusted p value for Holm-Bonferroni,
 Bonferroni, Dunnett's Test and Tukey's test. If compareall is TRUE, Dunnett's Test adjusted p.values
 will not show up. We can have the Tukey's test adjusted p value only if compareall is TRUE.
 "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" adjustment applied by p.adjust from
 p.adjust package.
}
\examples{
p.adj.compare(poisons$time, poisons$treat)
p.adj.compare(poisons$time, poisons$treat, control="C")
p.adj.compare(poisons$time, poisons$treat, compareall=TRUE, paired=TRUE)
}
\keyword{compare}
\keyword{groups}
