% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.2.R
\name{pairplot.2}
\alias{pairplot.2}
\title{Paired visualiztion of the data summary}
\usage{
pairplot.2(x1, x2, y = as.vector(cbind(x1, x2)), 
                  linecol = c(5:(length(unique(pair_id)) + 4)), 
                  main = "PAIRED PLOT", titlesize = 13, titleface = 'bold', 
                  titlefont = 'Times New Roman', titlecol = 1, titlehjust = 0.5, 
                  subtitle = NULL, subsize = 10, subface = NULL, 
                  subfont = 'Times New Roman', subcol = 1, subhjust = 0.5, 
                  caption = NULL, capsize = 9, capfont = 'Times New Roman', 
                  capcol = "grey48", caphjust = 0, xname = NULL, xnamesize = 12, 
                  xnamecol = 1, xnameface = 'bold', xnamefont = 'Times New Roman', 
                  xlabels = unique(x), xlabelsize = 10, xlabelcol = "grey48",
                  xlabelface = NULL, xlabelfont = 'Arial', yname = NULL, 
                  ynamesize = 12, ynamecol = 1, ynameface = 'bold',  yname = NULL,
                  ynamefont = 'Times New Roman', ylabelcol = "grey48", 
                  ylabelface = NULL, ylabelfont = 'Arial', legname = NULL, 
                  legcol = 1, legface = 'bold', legfont = 'Times New Roman', 
                  leglab = unique(pair_id), legtxsize = 9, legtxcol = 1, legsize = 15, 
                  legtxface = NULL, legtxfont = 'Times New Roman', legpos = "right", 
                  globalp_pos = max(y, na.rm = T), psize = 4, pfont = 'Times New Roman', 
                  na.rm = T, ...)
}
\arguments{
\item{x1}{First group values (paired to x2)}

\item{x2}{Second group values (paired to x1)}

\item{linecol}{The line color}

\item{main}{Title of the boxplot}

\item{titlesize}{Font size of title}

\item{titleface}{Font style of title (e.g. 'bold')}

\item{titlefont}{Font of title (e.g. 'Times New Roman')}

\item{titlecol}{Color of title (default is black)}

\item{titlehjust}{Position of title (default is middle)}

\item{subtitle}{Subtitle of the boxplot}

\item{subsize}{Font size of subtitle}

\item{subface}{Font style of the subtitle}

\item{subfont}{Font of subtitle}

\item{subcol}{Color of subtitle (default is black)}

\item{subhjust}{Position of subtitle (default is middle)}

\item{caption}{Caption of the boxplot}

\item{capsize}{Font size of caption}

\item{capfont}{Font of caption}

\item{capcol}{Color of caption (default is 'gray48')}

\item{caphjust}{Position of caption (default is left)}

\item{xname}{Name of the variable on the x-asis}

\item{xnamesize}{Font size of the variable on the x-asis}

\item{xnamecol}{Font color of the variable on the x-asis}

\item{xnameface}{Font style of the variable on the x-asis}

\item{xnamefont}{Font of the variable on the x-asis}

\item{xlabels}{Labels of the levels on the x-asis}

\item{xlabelsize}{Font size of the levels on the x-asis}

\item{xlabelcol}{Font color of the levels on the x-asis}

\item{xlabelface}{Font style of the levels on the x-asis}

\item{xlabelfont}{Font of the levels on the x-asis}

\item{yname}{Name of the variable on the y-asis}

\item{ynamesize}{Font size of the variable on the y-asis}

\item{ynamecol}{Font color of the variable on the y-asis}

\item{ynameface}{Font style of the variable on the y-asis}

\item{ynamefont}{Font of the variable on the y-asis}

\item{ylabelsize}{Labels of the levels on the y-asis}

\item{ylabelcol}{Font color of the levels on the y-asis}

\item{ylabelface}{Font style of the levels on the y-asis}

\item{ylabelfont}{Font of the levels on the y-asis}

\item{legname}{Name of legend when TWO predictors}

\item{legsize}{Font size of legend title}

\item{legcol}{Color of legend title (default is black)}

\item{legface}{Font style of legend title}

\item{legfont}{Font of legend title}

\item{leglab}{Labels of each level in legend}

\item{legtxsize}{Font size of legend levels}

\item{legtxcol}{Font color of legend levels}

\item{legtxface}{Font style of legend levels}

\item{legtxfont}{Font of legend levels}

\item{legpos}{Position of legend (default is right)}

\item{psize}{Text size of p-value}

\item{pfont}{Text font of p-value}

\item{globalp_pos}{The position of p-value}

\item{na.rm}{Whether to include missing values (default is TRUE)}

\item{...}{Further arguments to be passed to or from methods.}
}
\description{
Summarize continuous response variable by categorical explanatory variable with two sample paired plot
}
\details{
A Two Sample Paired Function

The pairplot.2 function return the two-sample paired plot from ggplot2. Values from both groups 
         should be continuous. The sample sizes from two groups should be matched and the data in two groups 
         should be matched. The p-value comes from t.test with paired = T.
}
\examples{
pairplot.2(week1, week12, linecol = c('lightblue','deepskyblue2','deepskyblue4',
           'blue','darkblue'), main = "PAIRED PLOT OF FOOD INTAKE", 
           titlesize = 17, titlecol = 'dodgerblue4', xname = 'WEEK', xnamesize = 14,
           xlabels = c('1','12'), yname = 'FOOD INTAKE (gm)', ynamesize = 14, 
           legname = 'Pair ID', leglab = c('Pair 1','Pair 2','Pair 3','Pair 4','Pair 5'),
           subtitle = 'Paired graph of two sample t test', subsize = 13, subface = 'bold', 
           subfont = 'Times New Roman', subcol = 'grey48', capsize = 9, capfont = 'Times New Roman', 
           caption = 'Note: Data from Angelo Bisazza (1978) Development of Aggressive Behaviourin 
           the Mouse (Mus Musculus):\\nEffects of Different Enviromental Conditions, Bolletino di 
           zoologia, 45:4, 353-357,DOI: 10.1080/11250007809440142', globalp_pos = 48,
           psize = 5)
}
\keyword{data}
\keyword{paired}
\keyword{summary}
\keyword{t.test}
